# gateway/docs/swagger.yaml
basePath: /
definitions:
  handlers.AuthRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  handlers.AuthResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/handlers.UserResponse'
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.Product:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      updatedAt:
        type: string
    type: object
  handlers.UserResponse:
    properties:
      id:
        type: string
      is_admin:
        type: boolean
      login:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Проксі API для auth та product сервісів
  title: Gateway API
  version: "1.0"
paths:
  /api/auth:
    post:
      consumes:
      - application/json
      description: Приймає логін і пароль, перевіряє користувача та повертає JWT токен
      parameters:
      - description: Дані користувача
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Некоректний запит
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Невірні дані
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Помилка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Авторизація користувача
      tags:
      - auth
  /api/product:
    delete:
      consumes:
      - application/json
      description: Проксі-ендпоінт, який передає запити у product-service.
      parameters:
      - description: ID продукту (для GET, PUT, DELETE)
        in: query
        name: id
        type: string
      - description: Кількість продуктів (для GET, за замовчуванням 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Product'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.Product'
        "400":
          description: Некоректний запит
          schema:
            type: string
        "404":
          description: Продукт не знайдено
          schema:
            type: string
        "500":
          description: Помилка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Операції з продуктами (проксі)
      tags:
      - product
    get:
      consumes:
      - application/json
      description: Проксі-ендпоінт, який передає запити у product-service.
      parameters:
      - description: ID продукту (для GET, PUT, DELETE)
        in: query
        name: id
        type: string
      - description: Кількість продуктів (для GET, за замовчуванням 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Product'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.Product'
        "400":
          description: Некоректний запит
          schema:
            type: string
        "404":
          description: Продукт не знайдено
          schema:
            type: string
        "500":
          description: Помилка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Операції з продуктами (проксі)
      tags:
      - product
    post:
      consumes:
      - application/json
      description: Проксі-ендпоінт, який передає запити у product-service.
      parameters:
      - description: ID продукту (для GET, PUT, DELETE)
        in: query
        name: id
        type: string
      - description: Кількість продуктів (для GET, за замовчуванням 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Product'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.Product'
        "400":
          description: Некоректний запит
          schema:
            type: string
        "404":
          description: Продукт не знайдено
          schema:
            type: string
        "500":
          description: Помилка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Операції з продуктами (проксі)
      tags:
      - product
    put:
      consumes:
      - application/json
      description: Проксі-ендпоінт, який передає запити у product-service.
      parameters:
      - description: ID продукту (для GET, PUT, DELETE)
        in: query
        name: id
        type: string
      - description: Кількість продуктів (для GET, за замовчуванням 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Product'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.Product'
        "400":
          description: Некоректний запит
          schema:
            type: string
        "404":
          description: Продукт не знайдено
          schema:
            type: string
        "500":
          description: Помилка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Операції з продуктами (проксі)
      tags:
      - product
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"